requirements.txt
apache-airflow==2.9.3
requests
psycopg2-binary
